refactor(notification): change directory of notification feature
