+++ backend/src/main/java/com/brokendev/backend/boleto_payment/service/BoletoPaymentService.java
@@ -10,7 +10,7 @@ import com.brokendev.backend.common.exceptions.InvalidBoletoAmountException;
 import com.brokendev.backend.common.exceptions.UserAccountNotFoundException;
 import com.brokendev.backend.common.exceptions.InsufficientBalanceException;
 import com.brokendev.backend.enums.BoletoPaymentStatus;
-import com.brokendev.backend.services.NotificationService;
+import com.brokendev.backend.notification.service.NotificationService;
 import org.springframework.stereotype.Service;
 
 import java.math.BigDecimal;
