+++ backend/src/main/java/com/brokendev/backend/notification/service/NotificationService.java
@@ -1,11 +1,11 @@
-package com.brokendev.backend.services;
+package com.brokendev.backend.notification.service;
 
-import com.brokendev.backend.domain.Notification;
+import com.brokendev.backend.notification.domain.Notification;
 import com.brokendev.backend.common.domain.user.User;
-import com.brokendev.backend.dto.notification.NotificationResponseDTO;
+import com.brokendev.backend.notification.dto.NotificationResponseDTO;
 import com.brokendev.backend.common.exceptions.NotificationAccessDeniedException;
 import com.brokendev.backend.common.exceptions.NotificationNotFoundException;
-import com.brokendev.backend.repositories.NotificationRepository;
+import com.brokendev.backend.notification.domain.NotificationRepository;
 import org.springframework.stereotype.Service;
 
 import java.time.LocalDateTime;
