+++ backend/src/test/java/com/brokendev/backend/pix_transfer/service/PixTransferServiceTest.java
@@ -12,7 +12,7 @@ import com.brokendev.backend.pix_transfer.domain.PixTransaction;
 import com.brokendev.backend.pix_transfer.domain.PixTransactionRepository;
 import com.brokendev.backend.pix_transfer.dto.PixTransferRequestDTO;
 import com.brokendev.backend.pix_transfer.dto.PixTransferResponseDTO;
-import com.brokendev.backend.services.NotificationService;
+import com.brokendev.backend.notification.service.NotificationService;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
